Flatten input module
